# Conflicts:
#	UI/src/app/services/configuration.service.ts

+++ database/load_csv.sh
@@ -3,5 +3,12 @@ apt-get install unzip wget -y
 
 cd tmp
 wget -O vocabulary.zip $VOC_URL
-unzip 'vocabulary.zip' -d vocabulary
-rm vocabulary.zip
+
+# Archive not empty
+if zipinfo vocabulary.zip > /dev/null; then
+	unzip 'vocabulary.zip' -d vocabulary
+	rm vocabulary.zip
+else
+	echo "Vocabulary link expired"
+	exit 1
+fi
